+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.17.3
 * `CHG` intelli-scense: treat `V[]` as `table<integer, V>` in `pairs`
 * `FIX` [#435](https://github.com/sumneko/lua-language-server/issues/435)
+* `FIX` [#436](https://github.com/sumneko/lua-language-server/issues/436)
 
 ## 1.17.2
 `2021-3-2`