+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.17.3
 * `CHG` intelli-scense: treat `V[]` as `table<integer, V>` in `pairs`
+* `FIX` completion: `detail` disappears during continuous input
 * `FIX` [#435](https://github.com/sumneko/lua-language-server/issues/435)
 * `FIX` [#436](https://github.com/sumneko/lua-language-server/issues/436)