+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 1.17.3
+`2021-3-3`
 * `CHG` intelli-scense: treat `V[]` as `table<integer, V>` in `pairs`
 * `FIX` completion: `detail` disappears during continuous input
 * `FIX` [#435](https://github.com/sumneko/lua-language-server/issues/435)